swarms-master-rustified/swarms/agents/create_agents_from_yaml_rustified.rs
# Viable for conversion: Yes, but requires careful handling of Python libraries and data structures equivalent in Rust.
# 
# The conversion will require the following steps:
# 1. Replace Pydantic with a similar Rust library for building robust and scalable data models.
# 2. Find Rust equivalents for libraries like `yaml`, `tenacity`, and `loguru_logger`.
# 3. Replace Python's `typing` module with Rust's type system.
# 4. Port the retry logic and exception handling to Rust.

Here is the Rust version of the provided Python code. Note that this example assumes the existence of certain libraries and data structures (like `Agent`, `SwarmRouter`, and `LiteLLM`) which would need to be defined in Rust.

```rust
use anyhow::{Context, Result};
use log::{info, error};
use serde::{Deserialize, Serialize};
use serde_yaml::{from_str, from_reader};
use std::collections::HashMap;
use std::fs::File;
use std::time::Duration;

// Define AgentConfig struct with Serialize and Deserialize traits
#[derive(Serialize, Deserialize)]
struct AgentConfig {
    agent_name: String,
    system_prompt: String,
    model_name: Option<String>,
    max_loops: i32,
    autosave: bool,
    dashboard: bool,
    verbose: bool,
    dynamic_temperature_enabled: bool,
    saved_state_path: Option<String>,
    user_name: String,
    retry_attempts: i32,
    context_length: i32,
    return_step_meta: bool,
    output_type: String,
    auto_generate_prompt: bool,
    artifacts_on: bool,
    artifacts_file_extension: String,
    artifacts_output_path: String,
}

// Define SwarmConfig struct with Serialize and Deserialize traits
#[derive(Serialize, Deserialize)]
struct SwarmConfig {
    name: String,
    description: String,
    max_loops: i32,
    swarm_type: String,
    task: Option<String>,
    flow: Option<HashMap<String, String>>,
    autosave: bool,
    return_json: bool,
    rules: String,
}

// Define YAMLConfig struct with Serialize and Deserialize traits
#[derive(Serialize, Deserialize)]
struct YAMLConfig {
    agents: Vec<AgentConfig>,
    swarm_architecture: Option<SwarmConfig>,
}

// Define a function for loading and validating YAML configuration
fn load_yaml_safely(yaml_file: &str, yaml_string: Option<String>) -> Result<YAMLConfig> {
    if let Some(yaml_string) = yaml_string {
        let config: YAMLConfig = from_str(&yaml_string).context("Failed to parse YAML string")?;
        Ok(config)
    } else if let Ok(file) = File::open(yaml_file) {
        let config: YAMLConfig = from_reader(file).context("Failed to parse YAML file")?;
        Ok(config)
    } else {
        Err(anyhow::anyhow!("Failed to open YAML file"))
    }
}

// Define a function for creating an agent with retry logic
#[retry(3, 500, std::time::Duration::from_millis(500))]
fn create_agent_with_retry(agent_config: &AgentConfig, model: &LiteLLM) -> Result<Agent> {
    // Create an agent based on the provided configuration
    Agent::new(
        &agent_config.agent_name,
        &agent_config.system_prompt,
        model,
        agent_config.max_loops,
        agent_config.autosave,
        agent_config.dashboard,
        agent_config.verbose,
        agent_config.dynamic_temperature_enabled,
        agent_config.saved_state_path.as_ref(),
        &agent_config.user_name,
        agent_config.retry_attempts,
        agent_config.context_length,
        agent_config.return_step_meta,
        &agent_config.output_type,
        agent_config.auto_generate_prompt,
        agent_config.artifacts_on,
        &agent_config.artifacts_file_extension,
        &agent_config.artifacts_output_path,
    )
}

// Define a function for creating agents and/or SwarmRouter based on YAML configurations
fn create_agents_from_yaml(
    yaml_file: &str,
    yaml_string: Option<String>,
    return_type: &str,
) -> Result<(Option<SwarmRouter>, Vec<Agent>)> {
    let config = load_yaml_safely(yaml_file, yaml_string)?;
    let mut agents = Vec::new();
    let mut swarm_router = None;

    for agent_config in &config.agents {
        let model_name = agent_config.model_name.as_ref().unwrap_or(&"gpt-4o".to_string());
        let model = LiteLLM::new(model_name);

        let agent = create_agent_with_retry(agent_config, &model)?;
        agents.push(agent);
    }

    if let Some(swarm_config) = &config.swarm_architecture {
        let mut swarm_agents = agents.clone();
        swarm_router = Some(SwarmRouter::new(
            &swarm_config.name,
            &swarm_config.description,
            swarm_config.max_loops,
            &mut swarm_agents,
            &swarm_config.swarm_type,
            swarm_config.task.as_ref(),
            swarm_config.flow.as_ref(),
            swarm_config.autosave,
            swarm_config.return_json,
            &swarm_config.rules,
        ));
    }

    match return_type {
        "auto" => {
            if let Some(router) = &swarm_router {
                info!("Created SwarmRouter with auto return type");
                Ok((Some(router.clone()), Vec::new()))
            } else {
                info!("Created agents with auto return type");
                Ok((None, agents))
            }
        }
        "swarm" => {
            if let Some(router) = &swarm_router {
                info!("Created SwarmRouter with swarm return type");
                Ok((Some(router.clone()), Vec::new()))
            } else {
                Err(anyhow::anyhow!("SwarmRouter not created"))
            }
        }
        "agents" => {
            info!("Created agents with agents return type");
            Ok((None, agents))
        }
        "both" => {
            if let Some(router) = &swarm_router {
                info!("Created SwarmRouter and agents with both return type");
                Ok((Some(router.clone()), agents))
            } else {
                Err(anyhow::anyhow!("SwarmRouter not created"))
            }
        }
        _ => Err(anyhow::anyhow!("Invalid return type")),
    }
}

fn main() -> Result<()> {
    env_logger::init();
    let result = create_agents_from_yaml("agents.yaml", None, "auto")?;
    info!("Created agents and/or SwarmRouter: {:?}", result);
    Ok(())
}
```

Please note that some parts of the original code may not be directly portable to Rust due to differences in the languages and their standard libraries. For instance, Pydantic's validation and the `tenacity` library do not have direct Rust equivalents, so you might need to use a combination of other libraries to achieve the same functionality.

Additionally, error handling has been replaced with `anyhow` crate, which is a popular error handling library for Rust.

Make sure to add the required dependencies to your `Cargo.toml`:

```toml
[dependencies]
anyhow = "1.0.54"
log = "0.4.14"
serde = { version = "1.0.130", features = ["derive"] }
serde_yaml = "0.8.23"
```

Replace the `Agent`, `SwarmRouter`, and `LiteLLM` types with your actual Rust implementations of these structs.

This example demonstrates how the original Python code can be adapted to Rust. However, this example is intended to serve as a starting point and may require adjustments to your specific use case and requirements.